fix wrong type for dtos in clickhandlers + comment call to ihm

+++ My project/Assets/Scripts/IHM/Board/CaseClickHandler.cs
@@ -36,7 +36,7 @@ public class CaseClickHandler : MonoBehaviour
         lastCaseClick = this;
         
         //valeur destPos ajouté au dto 
-        dto.destPos= (transform.position.x, transform.position.z);
+        dto.destPos= ((uint)transform.position.x, (uint)transform.position.z);
                
         //envoi du dto 
         plateau.SendDTO(dto);
